+++ packages/react/src/components/ChatInput/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ChatInput allows the user to build chat inputs for apps. It features internal components meant to build chat inputs with popovers, actions, and fields.
 
+## 2026-01-15 `1.0.1`
+
+#### 🎉 New features
+
+- Fixes types to adapt with React 18. ([#137](https://github.com/TiendaNube/nimbus-patterns/pull/137) by [@joacotornello](https://github.com/joacotornello))
+
 ## 2025-09-24 `1.0.0`
 
 #### 🎉 New features